feat(frontend): dedicated campaign tasks page replacing the settings tab
